Initial commit of a mem-mapped register class
The class is basic, but does include thread-safety via named (Posix) semaphores. The constructor allows the caller to specify whether the instance (for a given register address) is writable from a single caller in the entire system (always the first caller), or if all instances can write to the register. In all cases, all instances can read from the register.

Read/write is locked per thread, and these functions allow the caller to specify an infinite wait (default), a timed wait, or a non-blocking wait (via special timeout arg. value).