Simplify Register class by making all registers readable and writable from all instances
This simplification serves two purposes:
1. Eliminate problematic FILE_EXISTS issue for the "exclusive" writer, when a previous exclusive writer has crashed.
2. Simplify the constructor logic considerably.

The notion of read and write access makes the most sense if we are using a read/write lock, but POSIX doesn't support a named version, so we can't easily do so. Now the burden is on the Register class user to ensure that they are an appropriate writer.