Initial commit of the Time Sync Mgr node and related msgs/services
This node serves as an ROS proxy to the Chrony NTP service, allowing other nodes to add and remove NTP servers (by IP address), query the existing servers, and restore to the factory default configuration.

This is a good example of a Python node that proxies an underlying daemon service via it's config file (/etc/chrony/chrony.conf in this case). At some point it will be worth extracting the reusable features from this node (or possibly just expanding the node to be the single system service manager handling all similar services).