Add a "private" node handle to SDKNode class for namespace encapsulation
This allows us to selectively add params, services, topics, etc. to a namespace under the node rather than a global namespace. This is necessary for differentiating generic ROS elements in base classes. An example is the (soon-to-be-committed) TriggerableNode specialization of SDKNode, which subscribes to trigger-controller messages, but only wants to receive those meant for the subclass node (e.g., test pattern generator). There are other ways to solve the problem, but this is pretty elegant and low cost.