Introduce a TriggerInterface abstract class.
Concrete subclasses will be _contained_ in triggerable nodes, replacing the previous inheritance-based pattern, as this was getting complicated.

This class provides no implementation of the key methods, since these depend on whether or not the node leveraging this interface has register access to the FPGA's trigger subsystem, or is spoofing it in some way.