Implement complete nd_node namespace conventions
This involves building the namespace in the launch file from environment variables that include the rootname, device type, and device serial number. None of these components are included as part of the class constructors or data members any longer as it is all handled under the hood by the launcher.