Add NDNode::publishImage() method to push an image on the image transport topic(s)
The new method takes in a OpenCV cv::Mat pointer as the input image, our first ROS use of OpenCV. The real meat of this commit was establishing the correct package.xml and CMakeLists.txt changes to bring in OpenCV correctly. Note that by including OpenCV and cv_bridge as a "COMPONENTS" of package "catkin", we avoid having to explicitly provide include and link-library directives... we get them for free through things like catkin_INCLUDE_DIRS, etc.

Note that this method respects the _simulated_data parameter, supplying simulated data when required (as read from a specified filesystem location).