Move data save configuration from NDNodes into a separate save_data_interface
The new interface provides the save_data_prefix topic specified during RUI design and a specialized save_data_status notification topic for change-driven global notifications. Everything else remains the same from the external API perspective.

This is a reference design for an upcoming effort to collect functionality into more compartmentalized interface classes (to include ROS params, topics, services, etc. associated with that functionality). This was done to an extent previously with the TriggerInterface class, but this current implementation fleshes it out a bit more.

In the near future, all interface classes will derive from an abstract base class that provides most of the boilerplate. It is possible that the SDKNode class will be reduced to just a container class for these interfaces along with a run() method.