Introduce new base SDKInterface class to be subclassed by common interfaces
This class works in concert with the SDKNode class to provide a consistent and
partially-transparent method to setup all ROS IPC mechanisms for a node. This base
class should be subclassed as described in the sdk_interface.h header file.

Eventually quite a bit of SDKNode's current implementation will be replaced by a series of interfaces adhering to this new standard.