Add a trigger interface implementation for nodes with no connection to FPGA trigger implementation
This makes it possible to e.g., trigger the camera on a Jetson in exactly the same way as FPGA-linked sensors are triggered, albeit with additional latency. The interface subscribes directly to the sw_trigger topic, and on receipt of a trig_mask that ORs with it's own index, it will call a callback from the parent SDKNode.