Make SDKNode::init() a public method for better control over initialization timing.
Previously, init() was private and called only in SDKNode::run(). This was problematic for nodes that need earlier access to their parameters, and other ROS interface capabilities. Now init() can be called directly on a constructed instance of an SDKNode, though the original behavior of init() at the start of run() is preserved whenever init() was called beforehand (to avoid having to modify all the existing SDKNodes.

Another benefit is that subclasses can override virtual init() to do their own additional initialization.