Allow proper (configured) image names when advertising.
The image names are meant to be the names of the advertised topics, but we were previously advertising before retrieving names from the param server. This commit fixes that.

It is awkward, however, to retreive here in initPublishers (before retrieveParams gets called). Would be better to do retrieveParams first, but it publishes a message (status), so we can't do it that way.