Make autostart more generic
This commit changes the paths and some of the elements involed in the autostart process. It adds some necessary files (e.g., etc/sys_env.bash) and updates systemd service files and start-up shell scripts to allow this num_sdk_base implementation to work for all products.

The product differentiators are captured in the etc/sys_env.bash file, which is committed intentionally invalid. Trying to start the roslaunch systemd service without first editing this file for on the target will fail. This is by design, so that developers must specify the necessary environment variables in the sys_env file that establish the device type and serial number.

etc/roslaunch.service should be copied to /etc/systemd/system/ and systemd should be completely reloaded or rebooted.