drivers: roll LPP-based drivers into one
Because a pipeline node needs to import rospy AND opencv it is going to
consume at least 50MB of memory throughout its lifetime. This means that
we are limited to 10 such nodes, minus memory required for other nodes
and system processes. Here we turn 6 individual drivers into one
conglomerate. Both the design and the implementation are less clean, but
the 250MB of memory savings are worth it.