HACK: Force NDNode image files to be saved as BGR8
This is a last-minute change to get the encoding to work. For some reason, for proper display of ROS images (rqt_imageview, etc.), need to set the encoding to rgb8, but for OpenCV saving, need it to be bgr8 or else saved files have screwed up colors. The Gstreamer pipeline for GSMultiCam images outputs BGR, so it makes sense that something should be bgr8, but other NDNodes are bound to put out other encodings... this needs more investigation.