Fix some default params and save data prototype to be more universal
This catch-all commit mainly changes the prototype for SaveDataIfNecessary to take an ImgConstPtr rather than ImgPtr, as that promotion is safe in this function, and allows it to work directly from the recently committed new publishImage overload.

Also, default parameter values are set in a way that makes more sense based on their usage for ZedMultiCamNode (but appropriate for any 3dx device).