Ensure all NodeParams are known to param server
We force a call to retrieve() in the constructor of NodeParam. This guarantees that on a subsequent save_config message, all NodeParams are saved properly.

This change renders the entire retrieveParams() notion unnecessary. However, there are a few cases where logic besides calls to NodeParam::retrieve() happens inside retrieveParams() -- namely in Zynq nodes that use this as a place to update Register values -- so it will take some work to completely remove retrieveParams() from the codebase. There is no harm in leaving it for now except clutter.
retrieveParam() concept unnecessary