Issue warnings when defined NodeParams are not retrieved properly
The SDKNode implicitly requries that subclasses retrieve their parameters in an overridden retrieveParams() method... but that is easy to forget to do. This commit primarily provides a ROS warning when a parameter is not retrieved properly.

It paid dividends in early testing where it found an unretrieved node_3dX parameter (manual_resolution_enabled)... that is fixed in this commit, too.

And some whitespace changes that I am tired of fighting!