Use current timestamp instead of data timestamp in data filenames
(S3DX-20)

This commit fixes the major issue that the ROS clock for a node is initialized when that node starts, and doesn't respond to system clock updates thereafter. So for 3DX, with no RTC, it was impossible in normal operations to get valid timestamps in filenames. The trade here is that now the file timestamps correspond to the time the data is saved, not the time it was acquired.