Dhclient weirdness work-around
When releasing a DHCP IP address via dhclient -r, ALL IP addresses on that interface are deleted. This commit ensures that after that happens, we reaquire all configured addresses.

Note, addresses not known to the param server (i.e., added at runtime) will not be readded. We can fix that by storing them before the dhclient and then restoring them later if this becomes an issue, but seems like such a remote corner case that it isn't worth the effort right now.