Fix a bug that keeps DHCP from being disabled at runtime
Discovered this during debugging with HMI -- because the param server value for dhcp_enabled is not modified when the setting is updated via ROS (to preserve the "Reset" functionality), the subsequent set_ips_from_params() (bug fix from March 2021) overrides whatever DHCP setting was just applied. This commit skips the DHCP part of set_ips_from_params() in that case.