TimeSyncMgr: Inform the rest of the system on first NTP sync
This commit finishes the job of sending the sys_time_updated message when the system clock has changed. In this case, we are also sending it the first time we detect an NTP sync.

The main effect of sending the message is that the camera driver manager automatically restarts the camera driver whenever a time update occurs. I was seeing a problem with a late NTP sync during target localizer testing that led me down this path today.