NetworkManager: Initial wifi status and setup support
This commit brings in a rudimentary WiFi detection (based on presence of wlan0 in interface list) and access point start/stop.

This will be expanded later to include ability to connect to an external WiFi network (in addition to acting as an access point), and probably configurability of AP name and password.

Would also be good to display the number of AP clients as a security measure.