Clean out some AHRS driver code and build products
NE-92

This is part of a major revamp to nepi_edge_sdk_nav_pose. The ROS interface to nav/pose has been fully moved into that project. We are preserving the Lord AHRS code here (but disabling it in the build) as a future reference in case we need to redo that driver in the future.