nepi_autolauncher: New node to detect hardware and auto-launch nodes
For now this only supports hardware that is detectable by presence/absence of a file in the /dev filesystem. Initial implementation was tested against IQR pan/tilt (/dev/iqr_pan_tilt) and Pixhawk 6C (/dev/ttyUSB0).

The mapping from /dev file to node type is handled by config. file -- no effort is made to validate the hardware or this mapping.

It _should_ be possible to manage multiple versions of the same device in this way, but that is generally untested at this point due to a lack of redundant hardware.