Network Mgr: Auto-retry Wifi client connection on failure
This is a work-around to an observed issue where on power cycle (but not soft reboot) the Wifi client connection does not come up properly.

Now on each failure to connect to the Wifi network, a timer is started to retry in 3 seconds.

Also investigated delays between bringing up the interface and launching wpa_supplicant to no avail, but this auto-retry scheme seems to work consistently.