Add correct NTP sync status assessment logic
NE-138

Now we are using the chronyc sources 'mode' sync character to report the sync. status so that clients (e.g., RUI) don't have to rely on the time-since-last-contact field to assess the status.

All previous reported NTP status fields are preserved, so advanced tracking and logic can be done by the client app.