Support Jetson A/B scheme in addition to NEPI A/B scheme
For the Orin NX, which has no embedded flash, it makes more sense to use the Jetson A/B scheme and no NEPI Init rootfs.

This requires a second implementation of a few functions in sw_update_utils and a detection mechanism to tell when to use Jetson vs NEPI. Both are implemented here.

NOTE: This is not currently being used... only NEPI scheme is in effect.