Autolauncher: Simplify the config. file search when launching nodes
Now we simply look for a folder and config. file with the same exact name as the node that is being launched. This seems less flexible, but actually because of how our ConfigMgr works, this config file naming is the only one that will actually work with the rest of the system.

The launched node may have a unique id appended per auto-launch logic, in which case, that's the name that is used when searching for a config. file.