nepi_autolauncher: Handle mavlink and sealite autodetection, too
This is a big departure -- now there is an assumed distinction between "path devices" that are detected solely based on filesystem node existence and those that require more sophisticated protocol-level detection.

Mavlink (e.g., Pixhawk) and Sealite are the first two such instances but others can be added by
a) Defining a discovery function (in a separate file) that can be called to discover, start, and manage nodes via device-custom logic
b) Adding that discover call to the NEPIAutolauncher::detectAndManageDevices() function

The custom discovery functions must accept a list of off-limits devices to avoid manipulating anything that is already in use and return a modified list that includes any paths that they are now controlling.